Update: I should have re-watche devaslife's Neovim Configuration video again
There is an extension for telescope which gives you a fucking file
browser. But to be honest I am not going to use it ( just commented it
out in the code )... I am happy with 'mini.files'

I think I might even remove it completely... That's the purpose of
'telescope.lua.bak' file!