Update: Added my own version of Corne Keyboard Layout
I am going to try to use a layout without Home Row Modifiers and
also that only uses 34 keys... This is because I think my end-game is
going to consist of 34 keys.

This is the reason that I am uploading this version of my own keymap to
my dotfiles repository.